custom add slot support
add slot (oblong hole) support for custom pattern generation.
these are often used for mounting tabs of shells and power jacks.
used slotWidth and slotHeight in the pattern to define a slot.
the slot hole shape is up to the generator but rounded caps are
the industry standard.
circle shape indicate to use slot pads with rounded caps.