fix variation filtering
elements can have variations, add documented in
https://doc.qeda.org/core/component/#-variations-
the issue was that when generating, the variation specified when
the element was add (e.g. qeda add element@variation) was ignored.
this change fixes the issue.

WARNING:
when an element has variations, and no variation is specified,
the element is added in all variations.
but when generating, the variations overwrite each other.
no warning is output.
this issue has not been fixed (the generator does not know which
variation is generated to add this information in the file name,
...)