add config option to disable timestamp
qeda added a timestamp in the generated kicad libraries (symbols
and footprint).
this is not very versioning friendly since the files are changed
after generation although no symbols or footprints are.
the kicad library file format specifies a timestamp field.
the @nodate configuration option sets a static timestamp in this
case.