add setting to hide symbol pin icon
before when generating a symbol, an icon indicating the pin shape
of the footprint was added in the symbol next to the pin.
currently only the connector symbol used this feature.
starting with KiCad 6, this shape would overlap with the pin name.
not setting symbol.pinIcon to false in the settings will prevent
this shape from being drawn in the symbol.