Add support for second level variations
With this patch, it is possible to partially override e.g. housing parameters per variation.
Before, this was only possible by creating a separate housing node for each variation and
copying all relevant information into it. This made it very inconvenient to describe e.g. a
whole connector family in a single file.

An (incomplete) example of how this can be used:
```
housing:
  rowCount: 1
  rowCount@2: 2
  rowCount@3: 3
```