Add support for custom chamfer in Kicad 7
Chamfers can be added by setting either chamfer, chamferTopLeft, chamferTopRight, chamferBotLeft or chamferBotRight in the housing section.
E.g. to add a chamfer on the top left of pads 1, 2, 3 and 6, add 'chamferTopLeft: 1-3, 6'.
When chamfer is set, all sides of the listed pins will get a chamfer.
Chamfer sizes cannot be adjusted at the moment.