Allow for extra option-keys
With this patch, it is possible to define e.g. schematic.options1 to define some extra options
without removing the ones defined in the component. Main use for this are base components that
have some options defined, where the component based needs to add one or two without affecting
the set already defined in the base.

Signed-off-by: Matthias Breithaupt <m.breithaupt@vogl-electronic.com>