Kicad7: Add manufacturer output
There are now three ways to define the manufacturer (the field is currently only output for Kicad7):
1. By setting the manufacturer in the component yaml (i.e. 'manufacturer: Manufacturer 1, Co.')
2. By using the path name (in qeda.yaml, set config.manufacturer.source to 'path')
3. By using a README.md file in a directory (config.manufacturer.source set to 'readme')

Assuming a directory structure like this, the following happens:
library/level0/level1/level2/mycomponent.yaml
If source is set to 'path', usually, the name of the directory in level0 would be used as manufacturer name.
To change this behavior, you can set config.manufacturer.level to any integer value, representing the level as
above. If the level is were to be set to 10 in the above example, default behavior would be that the manufacturer
is not set. By setting config.manufacturer.clamplevel to true (actually, anything at the moment), the highest available
directory name would be used instead (so in the above example, a level of 10 would still result in manufacturer 'level2').

By setting source to 'readme' instead, the software searches for a file called 'README.md' at the selected level
(0 by default).

If a manufacturer is set in the component yaml, this usually takes precedence over the one gotten from path or README.
To change this, config.manufacturer.override can be set. This way, the manufacturer names from the path are used instead.

Signed-off-by: Matthias Breithaupt <m.breithaupt@vogl-electronic.com>