Add support for overrides
Overrides are very similar to config options with the difference that they are only
applied to components selected by the user.

Overrides can be defined in the config. E.g.
```yaml
config:
  overrides:
    myoverride:
      housing:
        options: polarized
```

Any defined override can be used by adding a tilde followed by the name
of the override to the library entry. E.g.
```yaml
library:
  - my/component_a~myoverride
  - my/component_b@variation~myoverride
```

Anything defined in the override will take precedence over the values set in the
component or by inheritance.

Signed-off-by: Matthias Breithaupt <m.breithaupt@vogl-electronic.com>